workaround bad interaction between the stub mechanism and flexmock
`flexmock` as of yet does not handle `pass_thru` calls when the
next method in the stack has been inserted after the creation
of the partial mock.

This is what happens in this instance, where Syskit's stub mechanism
(which the canbus test relies on) inserts the stubbed `watch` method
just before calling `configure`. Since the mock was already setup,
calling pass_thru won't work and `watch` is never called. Since the
stubbed `watch` is what creates the ports, the configuration fails
(Syskit complains that dynamic ports that should have been created
have not been)

Change the test to instead rely on the stub storing the list of watched
IDs (which makes more sense from a testing perspective anyways).